requirements.txt
PyAutoGUI~=0.9.54
pynput~=1.7.6
Pillow~=9.5.0